+++ libraries/kernel/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @yuants/kernel
 
-This log was last generated on Thu, 26 Oct 2023 18:03:30 GMT and should not be manually modified.
+This log was last generated on Tue, 07 Nov 2023 08:14:53 GMT and should not be manually modified.
+
+## 0.4.0
+Tue, 07 Nov 2023 08:14:53 GMT
+
+### Minor changes
+
+- feat: support kernel dump & restore
 
 ## 0.3.0
 Thu, 26 Oct 2023 18:03:30 GMT